fixed text input 패키지 분리
