+++ sdg/src/main/java/com/shopl/sdg/component/text_input/SDGInput.kt
@@ -21,7 +21,6 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
@@ -54,121 +53,16 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.shopl.sdg_common.enums.Keyboard
-import com.shopl.sdg_common.enums.OutlineType
 import com.shopl.sdg_common.ext.clickable
 import com.shopl.sdg_common.foundation.SDGColor
 import com.shopl.sdg_common.ui.components.IOText
-import com.shopl.sdg_common.ui.components.IOTypeface
 import com.shopl.sdg_common.ui.components.TypefaceConfig
 import com.shopl.sdg_resource.R
 import java.text.DecimalFormat
 
-
-
-@Composable
-fun SDGFixedTextInput(
-    outlineType: OutlineType,
-    height: Dp = 104.dp,
-    input: String?,
-    hint: String,
-    inputState: InputState,
-    focusRequester: FocusRequester? = null,
-    backgroundColor: Color = SDGColor.Neutral0,
-    marginValues: PaddingValues = PaddingValues(),
-    maxLength: Int = Int.MAX_VALUE,
-    enableOnError: Boolean = false,
-    onInputChange: (String) -> Unit,
-) {
-    val scrollState = rememberScrollState()
-    val focusManager = LocalFocusManager.current
-    val isKeyboardOpen by keyboardAsState()
-
-    LaunchedEffect(key1 = isKeyboardOpen) {
-        if (isKeyboardOpen == Keyboard.Closed) {
-            focusManager.clearFocus()
-        }
-    }
-
-    val bgColor = when (inputState) {
-        InputState.Enable -> backgroundColor
-        InputState.Disable -> backgroundColor
-        is InputState.Error -> SDGColor.Red300_a10
-    }
-
-    val inputModifier = Modifier
-        .padding(marginValues)
-        .fillMaxWidth()
-        .height(height)
-        .background(
-            color = bgColor,
-            shape = RoundedCornerShape(12.dp)
-        )
-
-    val textStyle = TextStyle(
-        color = SDGColor.Neutral700,
-        fontSize = 16.sp,
-        fontFamily = IOTypeface.REGULAR.fontFamily,
-        letterSpacing = 0.sp,
-        lineHeight = 20.sp,
-    )
-
-    val lineHeight = with(LocalDensity.current) { 20.sp.toPx() }
-
-    LaunchedEffect(key1 = scrollState.maxValue) {
-        scrollState.animateScrollBy(lineHeight)
-    }
-
-    BasicTextField(
-        modifier = inputModifier
-            .then(
-                if (outlineType == OutlineType.OUTLINE) {
-                    Modifier.border(
-                        width = 1.dp,
-                        color = SDGColor.Neutral200,
-                        shape = RoundedCornerShape(12.dp)
-                    )
-                } else {
-                    Modifier
-                }
-            )
-            .then(
-                if (focusRequester != null) {
-                    Modifier.focusRequester(focusRequester)
-                } else {
-                    Modifier
-                }
-            )
-            .padding(12.dp)
-            .verticalScroll(scrollState),
-        value = input ?: "",
-        onValueChange = {
-            if (it.length <= maxLength) {
-                onInputChange(it)
-            }
-        },
-        enabled = when (inputState) {
-            InputState.Enable -> true
-            is InputState.Error -> enableOnError
-            else -> false
-        },
-        textStyle = textStyle,
-        cursorBrush = SolidColor(SDGColor.Neutral700),
-        decorationBox = { textField ->
-            textField()
-            if (input.isNullOrEmpty()) {
-                IOText(
-                    text = hint,
-                    textColor = SDGColor.Neutral300,
-                    fontSize = 16.sp
-                )
-            }
-        },
-    )
-}
 
 /**
  * Copy & Paste deny
@@ -759,14 +653,7 @@ private fun PrevInput(
                 .fillMaxSize()
                 .background(color = SDGColor.Neutral0)
         ) {
-            SDGFixedTextInput(
-                marginValues = PaddingValues(20.dp),
-                outlineType = OutlineType.OUTLINE,
-                input = "12341234",
-                hint = "asdfasdf",
-                inputState = InputState.Enable,
-                onInputChange = {}
-            )
+
 
             SDGUnderlineInput(
                 marginValues = PaddingValues(20.dp),
